mainframe-services/pyproject.toml
[tool.poetry]
name = "mainframe-services"
version = "0.1.0"
description = "Backend service for Mainframe Studio"
authors = ["FPT Software AI Center <support.aic@fpt.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.2"
uvicorn = "^0.32.0"
pydantic = {extras = ["email"], version = "^2.9.2"}
pydantic-settings = "^2.6.0"
python-dotenv = "^1.0.1"
passlib = "^1.7.4"
python-jose = "^3.3.0"
python-multipart = "^0.0.12"
motor = "^3.6.0"
azure-storage-blob = "^12.23.1"
loguru = "^0.7.2"
gitpython = "^3.1.43"
antlr4-python3-runtime = "^4.13.2"
pygments-ibm-cobol-lexer = "^1.2"
saq = {extras = ["redis"], version = "^0.18.1"}
sqlglot = {extras = ["rs"], version = "^25.27.0"}
beautifulsoup4 = "^4.12.3"
azurebatchload = "^0.6.3"
charset-normalizer = "^3.4.0"
langchain = "^0.3.4"
langchain-openai = "^0.2.3"
numba = "^0.60.0"
pandarallel = "^1.6.5"
networkx = "^3.4.2"
langgraph = "^0.2.45"
langchain-community = "^0.3.5"
websockets = "^13.1"
nebula3-python = "^3.8.3"
chardet = "^5.2.0"
pika = "^1.3.2"
pandas = "^2.2.0"
openpyxl = "^3.1.2"
bcrypt = "^4.3.0"

[tool.poetry.dev-dependencies]
pytest = "^8.3.3"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
